storage_service/requirements.txt
pymongo
pika==1.3.1
python-dotenv==0.21.0